refactor: 設定管理のメインプロセス集約とドメイン層のファイル分割
