+++ src/shared/settings.ts
@@ -1,5 +1,3 @@
-import { TAG_PLACEHOLDERS } from './constants/placeholders';
-
 /**
  * アプリケーションのユーザー設定の型定義。
  */
@@ -9,11 +7,3 @@ export interface AppSettings {
   /** トラック番号のパディング桁数 */
   trackNumberPadding: number;
 }
-
-/**
- * デフォルトの設定値。
- */
-export const DEFAULT_SETTINGS: AppSettings = {
-  renamePattern: `${TAG_PLACEHOLDERS.TRACK_NUMBER} - ${TAG_PLACEHOLDERS.TITLE}`,
-  trackNumberPadding: 2
-};
